+++ platform-operator/internal/operatorinit/run_operator.go
@@ -20,7 +20,9 @@ import (
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/mysqlcheck"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/namespacewatch"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/reconcile"
-	integrationcontroller "github.com/verrazzano/verrazzano/platform-operator/experimental/controllers/integration/singlemodule"
+	integrationcascade "github.com/verrazzano/verrazzano/platform-operator/experimental/controllers/integration/cascade"
+	integrationsingle "github.com/verrazzano/verrazzano/platform-operator/experimental/controllers/integration/single"
+
 	modulehandlerfactory "github.com/verrazzano/verrazzano/platform-operator/experimental/controllers/module/component-handler/factory"
 	verrazzancontroller "github.com/verrazzano/verrazzano/platform-operator/experimental/controllers/verrazzano"
 	"github.com/verrazzano/verrazzano/platform-operator/internal/config"
@@ -272,18 +274,16 @@ func initModuleControllers(log *zap.SugaredLogger, mgr controllerruntime.Manager
 // initModuleControllers creates the integration controllers
 func initIntegrationControllers(log *zap.SugaredLogger, mgr controllerruntime.Manager) error {
 	// single module integration controller
-	if err := integrationcontroller.InitController(integrationcontroller.IntegrationControllerConfig{
+	if err := integrationsingle.InitController(integrationsingle.IntegrationControllerConfig{
 		ControllerManager: mgr,
-		ModuleHandlerInfo: modulehandlerfactory.NewModuleHandlerInfo(),
 	}); err != nil {
 		log.Errorf("Failed to start the integration controller:%v", err)
 		return err
 	}
 
-	// all modules integration controller
-	if err := integrationcontroller.InitController(integrationcontroller.IntegrationControllerConfig{
+	// other modules integration controller
+	if err := integrationcascade.InitController(integrationcascade.IntegrationControllerConfig{
 		ControllerManager: mgr,
-		ModuleHandlerInfo: modulehandlerfactory.NewModuleHandlerInfo(),
 	}); err != nil {
 		log.Errorf("Failed to start the integration controller:%v", err)
 		return err
